Update Hitchhikers_Guide_to_the_Datalake.md
General update on sections of this document.
Changed MSI to MI.

P.S.:
Added a section for key vault and MI under the security section. Looking for contributors who are willing to work with me on this piece.
Added a section for HA-DR. Looking for contributors who are willing to work with me on this piece.